add search and filter to bank partners
