+++ packages/client/package.json
@@ -6,6 +6,7 @@
     "@apollo/client": "^3.6.9",
     "@rtgs-global/components": "^1.0.28556",
     "graphql": "^16.5.0",
+    "lodash.debounce": "^4.0.8",
     "prop-types": "^15.8.1",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
