+++ packages/client/src/app/pages/Banks/components/index.js
@@ -1,2 +1,3 @@
-export { default as BankPartnersTab } from './BankPartnersTab';
 export { default as AddBankPartnerTab } from './AddBankPartnerTab';
+export { default as BankPartnersList } from './BankPartnersList';
+export { default as BankPartnersTab } from './BankPartnersTab';
